# Conflicts:
#	Assets/_Core/1_Scenes/Levels/Special Levels/Tutorial.unity
#	Assets/_Core/5_Player/PlayerCombat.cs
#	Assets/_Core/5_Player/Prefab/Player.prefab
#	Assets/_Core/6_Characters/Enemies/Destructible.cs
#	Assets/_Core/6_Characters/Enemies/EnemyCombat.cs
#	Assets/_Core/6_Characters/Enemies/GroundEnemy/GroundEnemy.cs

+++ Assets/_Core/6_Characters/Enemies/EnemyCombat.cs
@@ -23,7 +23,7 @@ namespace _Core._6_Characters.Enemies
             enemyData = (EnemyData) data;
         }
 
-        public override void OnAttackHit(float damage)
+        public override void OnAttackHit(int damage)
         {
             base.OnAttackHit(damage);
             TakeKnockback();
